+++ src/Interfaces.ts
